requirements.txt
line-bot-sdk==3.5.0
fastapi
uvicorn[standard]
langchain==0.2.5
langchain[all]
openai==0.28.1
yfinance
pydantic
tiktoken
flask==2.2.5
flask[async]
lxml==4.9.1
matplotlib
gspread==5.7.2
oauth2client==4.1.3
mpl_finance
arrow
python-dotenv==0.21.0
duckduckgo-search
langchain_community